script/requirements.txt
# Python requirements for NOAH platform deployment
# Ansible and related tools (pinned for collection compatibility)
ansible-core>=2.15.0,<2.17.0
ansible>=8.0.0,<10.0.0

# Ansible collections dependencies
requests>=2.31.0
urllib3>=2.0.0

# Cloud provider SDK dependencies
requests-toolbelt>=1.0.0

# Kubernetes client
kubernetes>=28.1.0
pyyaml>=6.0.1

# Certificate generation
cryptography>=41.0.0

# Note: For Ansible kubernetes.core modules, also install system package:
# sudo apt install python3-kubernetes (on Ubuntu/Debian)
# This is required for kubernetes.core collection to work properly

# Security and encryption
cryptography>=41.0.0
bcrypt>=4.0.0

# SSH and networking
paramiko>=3.3.0
netaddr>=0.9.0

# Jinja2 templating
jinja2>=3.1.0

# General utilities
packaging>=23.0
setuptools>=68.0.0

# Rich console library for enhanced CLI experience (used by Python scripts)
rich>=13.7.0

# Click CLI framework (used by noah.py)
click>=8.1.0
